Partial Completion
Worked on several minigame rooms using placeholder images
