+++ WakeUpRPGUnity/Assets/Scripts/EndCanvas.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aaf9f7198d00b074588bab62f1358196
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
