+++ WakeUpRPGUnity/Assets/Scripts/MinigameItemFindController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0266a925267165b43afd41a1217d404e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
