+++ WakeUpRPGUnity/Assets/Scripts/NarrationScene.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ce1eca7cb801a944cb6a085d15a4c3d1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
