+++ WakeUpRPGUnity/Assets/Scripts/StartCanvas.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 13813fd809ca02943a14eab7a79b7f95
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
