+++ WakeUpRPGUnity/Library/StateCache/MainStageHierarchy/4d/4d2a36d9faf65734f0bbefb3ee623532.json
@@ -1,1 +1,1 @@
-﻿{"m_ScrollY":0.0,"m_ExpandedSceneGameObjectInstanceIDs":[],"m_LastClickedInstanceID":0,"m_OpenSceneGUIDs":[]}
+﻿{"m_ScrollY":0.0,"m_ExpandedSceneGameObjectInstanceIDs":[-10490,-10488,-10284,-1338,-12,21424,21430,21436,21442,21456,21472,21476,21502,21512],"m_LastClickedInstanceID":21502,"m_OpenSceneGUIDs":["469de812e2f91124a84ee44869e43c81"]}
