+++ WakeUpRPGUnity/Library/StateCache/SceneView/8c/8cd7c613bf844de3b80696e27a479d5e.json
@@ -1,1 +1,1 @@
-﻿{"cameraMode":{"drawMode":3,"name":"Shadow Cascades","section":"Miscellaneous"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"m_AlwaysRefresh":false,"showFog":true,"showSkybox":false,"showFlares":true,"showImageEffects":true,"showParticleSystems":true,"showVisualEffectGraphs":true,"m_FxEnabled":true},"in2DMode":false,"pivot":{"x":258.68743896484377,"y":1337.519775390625,"z":-1800.97509765625},"rotation":{"x":0.3854886591434479,"y":0.04840511083602905,"z":-0.02025493048131466,"w":0.9212318658828735},"size":4.50320291519165,"orthographic":false}
+﻿{"cameraMode":{"drawMode":3,"name":"Shadow Cascades","section":"Miscellaneous"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"m_AlwaysRefresh":false,"showFog":true,"showSkybox":false,"showFlares":true,"showImageEffects":true,"showParticleSystems":true,"showVisualEffectGraphs":true,"m_FxEnabled":true},"in2DMode":false,"pivot":{"x":0.36031341552734377,"y":-1.467850685119629,"z":7.184597492218018},"rotation":{"x":-0.10649740695953369,"y":-0.04531565681099892,"z":0.0048605212941765789,"w":-0.9932923316955566},"size":5.736737251281738,"orthographic":false}
